requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-import-export==4.0.0
drf-spectacular==0.27.0
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
jinja2==3.1.3
django-jinja==2.11.0
celery==5.3.6
redis==5.0.1
django-cors-headers==4.3.1
whitenoise==6.6.0
gunicorn==21.2.0
